requirements.txt
fastapi
uvicorn[standard]
supabase
google-genai
python-dotenv
anyio
pydantic
